MAINT: Extract string helpers to a new private file
